# express-api/api-spec/halp_api_openapi_3.1.yaml
openapi: "3.1.0"
info:
    title: Halp API
    version: 0.1.0
...
paths:
...
    /auth/change-password:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
        patch:
            summary: update user password
            operationId: updatePasswordId
            tags:
                - auth
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PasswordChangeRequestSchema"
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/user:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
        get:
            summary: get logged in user info
            operationId: getUserInfoId
            tags:
                - auth
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
        post:
            summary: create team
            operationId: createTeamId
            tags:
                - auth
                - team
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TeamCreateRequestSchema"
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
        get:
            summary: get teams
            operationId: getTeamsId
            tags:
                - auth
                - team
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                contains:
                                    type: object
                                    properties:
                                        teams:
                                            type: array
                                            items:
                                                $ref: "#/components/schemas/TeamEnrichedSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
...
    /auth/teams/{teamId}/events:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
        post:
            summary: create event
            operationId: createEventId
            tags:
                - auth
                - event
                - admin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EventCreateRequestSchema"
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                    # TODO replace with actual object
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams/{teamId}/events/{eventId}:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
            - $ref: "#/components/parameters/EventIdParameter"
        get:
            summary: get managed event
            operationId: getEventManagedId
            tags:
                - auth
                - event
                - admin
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EventManagedSchema"
        delete:
            summary: delete event
            operationId: deleteEventId
            tags:
                - auth
                - event
                - admin
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams/{teamId}/events/{eventId}/jobs:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
            - $ref: "#/components/parameters/EventIdParameter"
        post:
            summary: create job
            operationId: createJobId
            tags:
                - auth
                - event
                - admin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/JobCreateRequestSchema"
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                    # TODO replace with actual object
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams/{teamId}/events/{eventId}/jobs/{jobId}:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
            - $ref: "#/components/parameters/EventIdParameter"
            - $ref: "#/components/parameters/JobIdParameter"
        delete:
            summary: delete job
            operationId: deleteJobId
            tags:
                - auth
                - event
                - admin
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
        patch:
            summary: reassign job
            operationId: reassignJobId
            tags:
                - auth
                - event
                - admin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/JobReassignRequestSchema"
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/events:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
        get:
            summary: get events
            operationId: getEventsId
            tags:
                - auth
                - event
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                contains:
                                    type: object
                                    properties:
                                        events:
                                            type: array
                                            items:
                                                $ref: "#/components/schemas/EventEnrichedSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/events/{eventId}:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/EventIdParameter"
        get:
            summary: get event
            operationId: getEventId
            tags:
                - auth
                - event
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/EventDetailedSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/events/{eventId}/volunteers:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/EventIdParameter"
        post:
            summary: volunteer for event
            operationId: volunteerForEventId
            tags:
                - auth
                - event
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                    # TODO replace with actual object
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
        delete:
            summary: unvolunteer from event
            operationId: unvolunteerFromEventId
            tags:
                - auth
                - event
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
components:
    parameters:
        AccessTokenParameter:
            name: Authorization
            in: header
            required: true
            schema:
                type: string
                # TODO add regex to match 'Bearer .+'
        TeamIdParameter:
            name: teamId
            in: path
            required: true
            schema:
                type: string
        EventIdParameter:
            name: eventId
            in: path
            required: true
            schema:
                type: string
...
        JobIdParameter:
            name: jobId
            in: path
            required: true
            schema:
                type: string
    responses:
        DefaultSuccessResponse:
            description: OK
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/DefaultSuccessResponseSchema"
        DefaultErrorResponse:
            description: Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/DefaultErrorResponseSchema"
    schemas:
...
        PasswordChangeRequestSchema:
            type: object
            properties:
                oldPassword:
                    type: string
                    format: password
                newPassword:
                    type: string
                    format: password
            required:
                - oldPassword
                - newPassword
        TeamCreateRequestSchema:
            type: object
            properties:
                teamName:
                    type: string
            required:
                - teamName
        EventCreateRequestSchema:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                dateTime:
                    type: number
                jobs:
                    type: array
                    minItems: 1
                    items:
                        $ref: "#/components/schemas/JobCreatorSchema"
            required:
                - name
                - dateTime
                - jobs
        JobCreateRequestSchema:
            type: object
            properties:
                jobType:
                    type: string
            required:
                - jobType
        JobReassignRequestSchema:
            type: object
            properties:
                volunteerId:
                    type: string

        # response body schemas
        DefaultSuccessResponseSchema:
            type: object
            properties:
                message:
                    type: string
            required:
                - message
        DefaultErrorResponseSchema:
            type: object
            properties:
                errorMessage:
                    type: string
            required:
                - errorMessage
...
        UserSchema:
            type: object
            properties:
                id:
                    type: string
                displayName:
                    type: string
                email:
                    type: string
            required:
                - id
                - displayName
                - email
        TeamEnrichedSchema:
            type: object
            properties:
                id:
                    type: string
                adminId:
                    type: string
                name:
                    type: string
                isSubscribed:
                    type: boolean
                isAdmin:
                    type: boolean
            required:
                - id
                - adminId
                - name
                - isSubscribed
                - isAdmin
        EventManagedSchema:
            type: object
            properties:
                event:
                    $ref: "#/components/schemas/EventSchema"
                volunteers:
                    type: array
                    items:
                        $ref: "#/components/schemas/VolunteeringEnrichedSchema"
                jobs:
                    type: array
                    minItems: 1
                    items:
                        $ref: "#/components/schemas/JobEnrichedSchema"
...
        TeamSchema:
            type: object
            properties:
                id:
                    type: string
                adminId:
                    type: string
                name:
                    type: string
        EventSchema:
            type: object
            properties:
                id:
                    type: string
                teamId:
                    type: string
                name:
                    type: string
                description:
                    type: string
                startDatetime:
                    type: number
                complete:
                    type: boolean
        EventEnrichedSchema:
            type: object
            properties:
                id:
                    type: string
                teamId:
                    type: string
                name:
                    type: string
                description:
                    type: string
                startDatetime:
                    type: number
                complete:
                    type: boolean
                isVolunteering:
                    type: boolean
                isAssigned:
                    type: boolean
            required:
                - id
                - teamId
                - name
                - description
                - startDatetime
                - complete
                - isVolunteering
                - isAdmin
        EventDetailedSchema:
            type: object
            properties:
                admin:
                    $ref: "#/components/schemas/UserSchema"
                team:
                    $ref: "#/components/schemas/TeamSchema"
                event:
                    $ref: "#/components/schemas/EventEnrichedSchema"
                jobs:
                    type: array
                    items:
                        $ref: "#/components/schemas/JobEnrichedSchema"
        VolunteeringEnrichedSchema:
            type: object
            properties:
                eventId:
                    type: string
                volunteer:
                    $ref: "#/components/schemas/UserSchema"
        JobCreatorSchema:
            type: object
            properties:
                jobName:
                    type: string
                count:
                    type: number
                    # TODO positive interger, ...
            required:
                - jobName
                - count
        JobEnrichedSchema:
            type: object
            properties:
                id:
                    type: string
                eventId:
                    type: string
                type:
                    type: string
                assignee:
                    $ref: "#/components/schemas/UserSchema"
            required:
                - id
                - eventId
                - type